refactor: move ensureWorkspacesArray function to workspace-resolver
